Supports installation via Gitee

+++ scripts/install.sh
@@ -18,9 +18,65 @@ sudo apt update
 sudo apt install dialog -y
 fi
 
+dpkg -s jq > /dev/null
+if [ $? != 0 ]; then
+sudo apt update
+sudo apt install jq -y
+fi
+
+
+rm -rf /tmp/._fts > /dev/null
+STATE=""
+dialog --clear --colors --backtitle "Fly-Tools ${LTAG} Installer (0/3)" --title "\Z2Select installation source" --menu "\Z4If you can access Github, please try to use Github to install it." 20 60 10 \
+"Github"    "Use Github to install it." \
+"Gitee"     "Use Gitee to install it." \
+"Exit"      "Exit the installer" \
+2>/tmp/._fts
+
+STATE=$(cat /tmp/._fts)
+rm -rf /tmp/._fts > /dev/null
+
+case $STATE in
+    Exit)
+        echo "\nGoodbye!!!\n"
+        exit 0
+        ;;
+    "")
+        echo "\nGoodbye!!!\n"
+        exit 0
+        ;;
+    "Github")
+        GIT_URL="github.com"
+        GIT_API="api.github.com"
+        GIT_USER="kluoyun"
+        GIT_TTYD_USER="tsl0922"
+    ;;
+    "Gitee")
+        GIT_URL="gitee.com"
+        GIT_API="gitee.com/api/v5"
+        GIT_USER="zxkxz"
+        GIT_TTYD_USER="ZXKXZ"
+    ;;
+esac
+
+
+LTAG=$(curl -Ls "https://${GIT_API}/repos/${GIT_USER}/FLY-Tools/releases/latest" | jq -r '.tag_name')
+if [ "$?" != "0" ]; then
+    dialog --colors --backtitle "Fly-Tools Installer (1/3)" --title "\Z1\ZbInstall Error Fly-Tools" --infobox "\Z3Unable to get latest version.\n If it is a gitee source, it may be that your installation times are too many, resulting in IP restrictions." 10 60
+    echo "\n"
+    exit 1
+fi
+TTYD_LTAG=$(curl -Ls "https://${GIT_API}/repos/${GIT_TTYD_USER}/ttyd/releases/latest" | jq -r '.tag_name')
+if [ "$?" != "0" ]; then
+    dialog --colors --backtitle "Fly-Tools Installer (1/3)" --title "\Z1\ZbInstall Error ttyd" --infobox "\Z3Unable to get latest version.\n If it is a gitee source, it may be that your installation times are too many, resulting in IP restrictions." 10 60
+    echo "\n"
+    exit 1
+fi
+
 M=$(uname -m)
 rm -rf /tmp/._fts > /dev/null
-dialog --clear --title "Fly-Tools Installer" --menu "The current system architecture is: ${M}\nChoose your system architecture." 20 60 10 \
+STATE=""
+dialog --clear --colors --backtitle "Fly-Tools ${LTAG} Installer (1/3)" --title "\Z2Choose system architecture" --menu "\Z4The current system architecture is: \Z1\Zr\Zb${M}\Zn \n\Z0Choose your system architecture." 20 60 10 \
 "linux-amd64"    "X86_64 ..." \
 "linux-arm64"    "aarch64, arm64-v8a arm64 ..." \
 "linux-arm"      "armeabi-v7a, armhf, arm ..." \
@@ -31,31 +87,33 @@ dialog --clear --title "Fly-Tools Installer" --menu "The current system architec
 "linux-mips64le" "Mips64le ..." \
 "linux-s390x"    "S390x ..." \
 "Other"          "Please install manually for other architectures" \
-"Exit" "Exit the installer" \
+"Exit"           "Exit the installer" \
 2>/tmp/._fts
 
 STATE=$(cat /tmp/._fts)
 rm -rf /tmp/._fts > /dev/null
-dialog --clear
 
 case $STATE in
     Exit)
+        echo "\nGoodbye!!!\n"
         exit 0
         ;;
     Other)
         echo "Please go to the release page to download the executable file and install it yourself."
+        echo "\nGoodbye!!!\n"
         exit 0
         ;;
     "")
+        echo "\nGoodbye!!!\n"
         exit 0
         ;;
     *)
-        LTAG=$(curl -Ls "https://api.github.com/repos/kluoyun/Fly-Tools/releases/latest" | grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/')
-        URL="https://github.com/kluoyun/Fly-Tools/releases/download/${LTAG}/Fly-Tools_${STATE}"
+        
+        URL="https://${GIT_URL}/${GIT_USER}/FLY-Tools/releases/download/${LTAG}/Fly-Tools_${STATE}"
 
         case $STATE in
             linux-amd64)
-                TTYD_STATE="X86_64"
+                TTYD_STATE="x86_64"
                 ;;
             linux-arm64)
                 TTYD_STATE="aarch64"
@@ -82,23 +140,34 @@ case $STATE in
                 TTYD_STATE="s390x"
                 ;;
         esac
-        TTYD_LTAG=$(curl -Ls "https://api.github.com/repos/tsl0922/ttyd/releases/latest" | grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/')
-        TTYD_URL="https://github.com/tsl0922/ttyd/releases/download/${TTYD_LTAG}/ttyd.${TTYD_STATE}"
+        TTYD_URL="https://${GIT_URL}/${GIT_TTYD_USER}/ttyd/releases/download/${TTYD_LTAG}/ttyd.${TTYD_STATE}"
         ;;
 esac
 
-echo $URL
-echo $TTYD_URL
+# echo FLY-Tools url: $URL
+# echo ttyd url: $TTYD_URL
+# echo ""
 
 sudo cat > /etc/default/ttyd << EOF
 # /etc/default/ttyd
 
-TTYD_OPTIONS="-p 9998 -6 -O login"
+TTYD_OPTIONS="--writable -p 9998 -6 -O login"
 EOF
 
 sudo systemctl stop ttyd-f.service > /dev/null 2>&1
 sudo rm /usr/local/bin/ttyd-f > /dev/null 2>&1
-sudo wget -c $TTYD_URL -O /usr/local/bin/ttyd-f #> /dev/null 2>&1
+
+wget --no-check-certificate -c -O ./ttyd-f $TTYD_URL 2>&1 | \
+stdbuf -o0 awk '/[.] +[0-9][0-9]?[0-9]?%/ { print substr($0,63,3) }'| \
+dialog --backtitle "Fly-Tools ${LTAG} Installer (2/3)" --title "Download components (1/2)" --gauge "Download ttyd-${TTYD_STATE}-${TTYD_LTAG}" 10 80
+
+if [ ! -f "./ttyd-f" ] || [ ! -s "./ttyd-f" ]; then
+    sudo rm ./ttyd-f > /dev/null 2>&1
+    dialog --colors --backtitle "Fly-Tools ${LTAG} Installer (2/3)" --title "\Z1Download Error\Zn" --infobox "Download ttyd-${TTYD_STATE}-${TTYD_LTAG} failed, please check the network" 10 60
+    echo "\n"
+    exit 1
+fi
+sudo mv ./ttyd-f /usr/local/bin/ttyd-f
 sudo chmod +x /usr/local/bin/ttyd-f
 
 sudo touch /etc/systemd/system/ttyd-f.service
@@ -124,6 +193,11 @@ EOF
 sudo systemctl daemon-reload
 sudo systemctl enable ttyd-f.service
 sudo systemctl restart ttyd-f.service
+if [ "$?" != "0" ]; then
+    dialog --colors --backtitle "Fly-Tools ${LTAG} Installer (2/3)" --title "\Z1Install ttyd Error\Zn" --infobox "The ttyd service failed to install and start successfully." 10 60
+    echo "\n"
+    exit 1
+fi
 
 if [ -e "/etc/systemd/system/FLY-Tools.service" ];then
     sudo systemctl stop FLY-Tools.service > /dev/null 2>&1
@@ -134,7 +208,18 @@ fi
 
 sudo systemctl stop Fly-Tools.service > /dev/null 2>&1
 sudo rm /usr/local/bin/Fly-Tools > /dev/null 2>&1
-sudo wget -c $URL -O /usr/local/bin/Fly-Tools #> /dev/null 2>&1
+
+wget --no-check-certificate -c -O ./Fly-Tools $URL 2>&1 | \
+stdbuf -o0 awk '/[.] +[0-9][0-9]?[0-9]?%/ { print substr($0,63,3) }'| \
+dialog --backtitle "Fly-Tools ${LTAG} Installer (2/3)" --title "Download components (2/2)" --gauge "Download Fly-Tools-${STATE}-${LTAG}" 10 80
+
+if [ ! -f "./Fly-Tools" ] || [ ! -s "./Fly-Tools" ]; then
+    sudo rm ./Fly-Tools > /dev/null 2>&1
+    dialog --colors --backtitle "Fly-Tools ${LTAG} Installer (2/3)" --title "\Z1Download Error\Zn" --infobox "Download Fly-Tools-${STATE}-${LTAG} failed, please check the network" 10 60
+    echo "\n"
+    exit 1
+fi
+sudo mv ./Fly-Tools /usr/local/bin/Fly-Tools
 sudo chmod +x /usr/local/bin/Fly-Tools
 
 sudo touch /etc/systemd/system/Fly-Tools.service
@@ -160,5 +245,13 @@ EOF
 sudo systemctl daemon-reload
 sudo systemctl restart Fly-Tools.service
 sudo systemctl enable Fly-Tools.service
+if [ "$?" != "0" ]; then
+    dialog --colors --backtitle "Fly-Tools ${LTAG} Installer (2/3)" --title "\Z1Install Fly-Tools Error\Zn" --infobox "The Fly-Tools service failed to install and start successfully." 10 60
+    echo "\n"
+    exit 1
+fi
 
-echo "安装结束"
+dialog --colors --backtitle "Fly-Tools ${LTAG} Installer (3/3)" --title "\Z4Fly-Tools installed successfully\Zn" --msgbox "Fly-Tools installation completed.\nPlease open your browser and visit http://{ip}:9999/\n\nPress the \Z1\ZrEnter\Zn key to exit." 10 60
+
+echo "\nGoodbye!!!\n"
+exit 0
